requirements.base.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Cython==0.29.32
docker-pycreds==0.4.0
filelock==3.9.0
gitdb==4.0.10
GitPython==3.1.30
huggingface-hub==0.12.0
idna==3.4
install==1.3.5
joblib==1.2.0
memory-profiler==0.61.0
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
sentencepiece==0.1.97
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
typing-extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
